Use TAP instead of TUN on linux
A TAP can have a bridge master, and we'll need that for the docker
network plugin.
